python_scraper/requirements.txt
requests==2.31.0
mysql-connector-python==8.4.0
python-dotenv==1.2.2
tqdm>=4.66.0